requirements.txt
accelerate @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_accelerate_1750456149/work
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.0
blinker==1.9.0
Bottleneck @ file:///croot/bottleneck_1731058641041/work
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
build==1.2.2.post1
cachetools==5.5.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1749972191589/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
chromadb==1.0.15
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
datasets==3.6.0
defusedxml==0.7.1
dill @ file:///croot/dill_1715094664823/work
distro==1.9.0
durationpy==0.10
executing==2.2.0
faiss-cpu==1.11.0
filelock @ file:///croot/filelock_1744281381737/work
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec @ file:///croot/fsspec_1750681162352/work
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///croot/gmpy2_1738085463648/work
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.23.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface_hub @ file:///croot/huggingface_hub_1747423383041/work
humanfriendly==10.0
icecream==2.1.5
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2 @ file:///croot/jinja2_1741710844255/work
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.26
langchain-core==0.3.66
langchain-google-genai==2.1.6
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langsmith==0.4.3
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
marshmallow==3.26.1
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mmh3==5.1.0
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
multidict==6.6.0
multiprocess @ file:///work/perseverance-python-buildout/croot/multiprocess_1698870874259/work
mypy_extensions==1.1.0
narwhals==1.45.0
networkx @ file:///croot/networkx_1737039604450/work
numexpr @ file:///croot/numexpr_1730215937391/work
numpy @ file:///croot/numpy_and_numpy_base_1747238018250/work/dist/numpy-2.2.5-cp312-cp312-linux_x86_64.whl#sha256=5ab22c86c5881c847fa1bb8b572e198c11ea7ebab63d5d24fbc9db9bf24d2365
oauthlib==3.3.1
onnxruntime==1.22.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging @ file:///croot/packaging_1734472117206/work
pandas @ file:///croot/pandas_1732735089971/work/dist/pandas-2.2.3-cp312-cp312-linux_x86_64.whl#sha256=57b66702d418720ec8483f7c4ec7c08d41815316ad7ce09d5b7bbc34eefcfdfd
pillow @ file:///croot/pillow_1744613067434/work
posthog==5.4.0
propcache==0.3.2
proto-plus==1.26.1
protobuf @ file:///croot/protobuf_1742419705443/work/bazel-bin/python/dist/protobuf-5.29.3-cp312-abi3-linux_x86_64.whl#sha256=b288453436273a46653983ade721c0c62cfa5ef6c012b110674c369d25536d88
psutil @ file:///croot/psutil_1736367091698/work
pyarrow @ file:///croot/pyarrow_1738229418212/work/python
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.1.1
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
referencing==0.36.2
regex @ file:///croot/regex_1736540786412/work
requests @ file:///croot/requests_1750426329888/work
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
safetensors @ file:///croot/safetensors_1741361308036/work
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==72.1.0
shellingham==1.5.4
six @ file:///croot/six_1744271502820/work
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
streamlit==1.46.1
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers @ file:///croot/tokenizers_1741370336077/work
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.5.1
tqdm @ file:///croot/tqdm_1738943501192/work
transformers @ file:///home/conda/feedstock_root/build_artifacts/transformers_1748607927906/work
triton==3.1.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
uritemplate==4.2.0
urllib3 @ file:///croot/urllib3_1750775463400/work
uvicorn==0.35.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
xxhash @ file:///croot/python-xxhash_1737039434400/work
yarl==1.20.1
youtube-transcript-api==1.1.0
yt-dlp==2025.6.25
zipp==3.23.0
zstandard==0.23.0
